+++ domain/src/main/java/com/riftar/domain/stockchart/model/ChartResult.kt
@@ -2,6 +2,6 @@ package com.riftar.domain.stockchart.model
 
 data class ChartResult(
     val meta: Meta,
-    val timestamp: List<Int>,
+    val timestamp: List<Long>,
     val indicators: Indicators
 )
